Armenia and Kar-Kiya patch
-	Added possibility for Armenia to break free If the Shaki fail their early wars. The Player can attempt to lead an uprising against the Shaki and play the newly added Armenian content which contains:
o	Flavour-events for the process of civilising
o	Armenian expansion and gaining cores in the Caucasus and Anatolia over the course of the game
o	An event chain regarding the state of the Armenian Cilicians
o	A decision to restore the Empire of Tigranes the Great for absolute Monarchies and fascist dictatorships and a less reactionary equivalent for socialist governments
o	Decisions to accept Assyrians, Georgians and Syriacs
o	Flavour events such as the reintegration of the Armenian diaspora
-	Added an event chain to the Kar-Kiya leading up to the constitutional revolution. The choices taken during the events unlock one of the three possible country paths for the Kar-Kiya
o	Constitutionalist – Nationalist. Encourage unification rebels in other Iranian countries, liberate the Caspian Iranians in Azerbaijan and reform the administration and economy of your realm, among other decisions.
o	Zaydi – Theocratic. Content regarding the persecution of the other Shi’a sects within your realm, expansion towards the holy shrines in Khorasan and the protection of the Zaydi minority in Yemen.
o	Royalist. Take advantage of the Russian patrons that helped keep the monarchy in power – at the risk of the Russians making unacceptable demands back. Possiblity of getting a new tech school and conflict if Russia decides to expand into the southern Caspian shore as the first step towards India.
-	Reduced the maluses Italy receives upon integrating Venice, Sardinia and Corsica
-	Reduced the militancy maluses the Sublime Porte receives when the Janissaries win the Civil War, preventing the depopulation of Bulgaria
-	Proclaiming the Kingdom of Hanover now gives Rhenish as an accepted culture
-	Taking the Baltic expansion path as Russia no longer leaves a few random provinces uncored
-	Mostar is now properly considered a port province, no longer making Illyria a landlocked nation
-	Fixed naval base in Christanborg, preventing bugged ship building
-	Fixed positions in Kiel
-	Fixed the Scramble for Africa full annexation CB not working as intended
-	The decision to integrate the Cape Coloured can no longer be taken as a Vryland-formed South Africa
-	The decision to integrate Greeks as a Venice who failed in the Council of Athens now properly works for both regular Republic and Merchant Republic governments
-	Various miscellaneous localisation fixes
-	Fixed improper promotion of serfs to farmers in mineral RGO provinces
-	Neu Kurland and Tobago now start as states, making Neu Kurland runs more viable
-	Eendrachtsland now has Burgundian rather than Bosnian ship names
-	Tarfaya, Aaiun and Dakhla no longer start as colonial provinces
-	Dissolving Gran Colombia as an Arcadian/Amerigan Great Power now properly releases Lusitania and Cuba